split paths/octoprint config into separate modules
