+++ services/src/config.rs
@@ -14,11 +14,9 @@ use serde_json::Value;
 
 use super::error::PrintNannyConfigError;
 use super::keys::PrintNannyKeys;
+use super::octoprint::OctoPrintConfig;
+use super::paths::{PrintNannyPaths, PRINTNANNY_CONFIG_DEFAULT};
 use printnanny_api_client::models;
-
-pub const OCTOPRINT_DIR: &str = "/home/octoprint/.octoprint";
-pub const PRINTNANNY_CONFIG_FILENAME: &str = "default.toml";
-pub const PRINTNANNY_CONFIG_DEFAULT: &str = "/etc/printnanny/default.toml";
 
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ArgEnum)]
 pub enum ConfigFormat {
@@ -184,62 +182,6 @@ impl Default for PrintNannyCloudProxy {
 }
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-pub struct PrintNannyPaths {
-    pub etc: PathBuf,
-    pub confd: PathBuf,
-    pub events_socket: PathBuf,
-    pub license: PathBuf,
-    pub issue_txt: PathBuf,
-    pub log: PathBuf,
-    pub octoprint: PathBuf,
-    pub run: PathBuf,
-    pub os_release: PathBuf,
-}
-
-impl Default for PrintNannyPaths {
-    fn default() -> Self {
-        // /etc is mounted as an r/w overlay fs
-        let etc: PathBuf = "/etc/printnanny".into();
-        let confd: PathBuf = "/etc/printnanny/conf.d".into();
-        let issue_txt: PathBuf = "/boot/issue.txt".into();
-        let run: PathBuf = "/var/run/printnanny".into();
-        let log: PathBuf = "/var/log/printnanny".into();
-        let events_socket = run.join("events.socket").into();
-        let license = "/boot/license.json".into();
-        let octoprint = OCTOPRINT_DIR.into();
-        let os_release = "/etc/os-release".into();
-        Self {
-            etc,
-            confd,
-            run,
-            issue_txt,
-            log,
-            events_socket,
-            octoprint,
-            license,
-            os_release,
-        }
-    }
-}
-
-impl PrintNannyPaths {
-    pub fn data(&self) -> PathBuf {
-        self.etc.join("data")
-    }
-    pub fn octoprint_venv(&self) -> PathBuf {
-        self.octoprint.join("venv")
-    }
-
-    pub fn octoprint_pip(&self) -> PathBuf {
-        self.octoprint_venv().join("bin/pip")
-    }
-
-    pub fn octoprint_python(&self) -> PathBuf {
-        self.octoprint_venv().join("bin/pip")
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct PrintNannyConfig {
     pub printnanny_cloud_proxy: PrintNannyCloudProxy,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -254,6 +196,7 @@ pub struct PrintNannyConfig {
     pub janus_edge_stream: Option<models::JanusEdgeStream>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub janus_cloud_stream: Option<models::JanusCloudStream>,
+    pub octoprint: OctoPrintConfig,
     pub paths: PrintNannyPaths,
     pub api: models::PrintNannyApiConfig,
     pub dash: DashConfig,
@@ -268,7 +211,7 @@ const FACTORY_RESET: [&'static str; 8] = [
     "device",
     "janus_edge",
     "janus_cloud",
-    "octoprint_install",
+    "octoprint_server",
     "user",
 ];
 
@@ -286,6 +229,7 @@ impl Default for PrintNannyConfig {
         let dash = DashConfig::default();
         let printnanny_cloud_proxy = PrintNannyCloudProxy::default();
         let keys = PrintNannyKeys::default();
+        let octoprint = OctoPrintConfig::default();
         PrintNannyConfig {
             api,
             dash,
@@ -293,6 +237,7 @@ impl Default for PrintNannyConfig {
             paths,
             printnanny_cloud_proxy,
             keys,
+            octoprint,
             alert_settings: None,
             cloudiot_device: None,
             device: None,
